+++ ubo_app/services/010-voice/reducer.py
@@ -1,6 +1,8 @@
 # ruff: noqa: D100, D101, D102, D103, D104, D107, N999
 from __future__ import annotations
 
+from dataclasses import replace
+
 from redux import CompleteReducerResult, InitAction, InitializationActionError
 
 from ubo_app.store.services.voice import (
@@ -9,6 +11,7 @@ from ubo_app.store.services.voice import (
     VoiceReadTextAction,
     VoiceState,
     VoiceSynthesizeTextEvent,
+    VoiceUpdateAccessKeyStatus,
 )
 
 
@@ -20,6 +23,9 @@ def reducer(
         if isinstance(action, InitAction):
             return VoiceState()
         raise InitializationActionError(action)
+
+    if isinstance(action, VoiceUpdateAccessKeyStatus):
+        return replace(state, is_access_key_set=action.is_access_key_set)
 
     if isinstance(action, VoiceReadTextAction):
         return CompleteReducerResult(